+++ Ergo/Interpreter/Libraries/CSharp/CSharp.cs
@@ -1,13 +1,14 @@
 ﻿using Ergo.Interpreter.Directives;
 using Ergo.Solver.BuiltIns;
 
-namespace Ergo.Interpreter.Libraries;
+namespace Ergo.Interpreter.Libraries.CSharp;
 
-public class Dict : Library
+public class CSharp : Library
 {
-    public override Atom Module => WellKnown.Modules.Dict;
+    public override Atom Module => WellKnown.Modules.CSharp;
     public override IEnumerable<SolverBuiltIn> GetExportedBuiltins() => Enumerable.Empty<SolverBuiltIn>()
-        .Append(new DictKeyValue())
+        .Append(new Pull())
+        .Append(new Push())
         ;
     public override IEnumerable<InterpreterDirective> GetExportedDirectives() => Enumerable.Empty<InterpreterDirective>()
         ;
