+++ Ergo/Interpreter/Libraries/IO/IO.cs
@@ -1,14 +1,15 @@
 ﻿using Ergo.Interpreter.Directives;
 using Ergo.Solver.BuiltIns;
 
-namespace Ergo.Interpreter.Libraries;
+namespace Ergo.Interpreter.Libraries.IO;
 
-public class List : Library
+public class IO : Library
 {
-    public override Atom Module => WellKnown.Modules.List;
+    public override Atom Module => WellKnown.Modules.IO;
     public override IEnumerable<SolverBuiltIn> GetExportedBuiltins() => Enumerable.Empty<SolverBuiltIn>()
-        .Append(new Nth0())
-        .Append(new Nth1())
+        .Append(new Write())
+        .Append(new WriteCanonical())
+        .Append(new WriteQuoted())
         ;
     public override IEnumerable<InterpreterDirective> GetExportedDirectives() => Enumerable.Empty<InterpreterDirective>()
         ;
