+++ Ergo/Interpreter/Libraries/Math/Math.cs
@@ -1,15 +1,14 @@
 ﻿using Ergo.Interpreter.Directives;
 using Ergo.Solver.BuiltIns;
 
-namespace Ergo.Interpreter.Libraries;
+namespace Ergo.Interpreter.Libraries.Math;
 
-public class IO : Library
+public class Math : Library
 {
-    public override Atom Module => WellKnown.Modules.IO;
+    public override Atom Module => WellKnown.Modules.Math;
     public override IEnumerable<SolverBuiltIn> GetExportedBuiltins() => Enumerable.Empty<SolverBuiltIn>()
-        .Append(new Write())
-        .Append(new WriteCanonical())
-        .Append(new WriteQuoted())
+        .Append(new Eval())
+        .Append(new Is())
         ;
     public override IEnumerable<InterpreterDirective> GetExportedDirectives() => Enumerable.Empty<InterpreterDirective>()
         ;
