+++ Ergo/Interpreter/Libraries/String/String.cs
@@ -1,7 +1,7 @@
 ﻿using Ergo.Interpreter.Directives;
 using Ergo.Solver.BuiltIns;
 
-namespace Ergo.Interpreter.Libraries;
+namespace Ergo.Interpreter.Libraries.String;
 
 public class String : Library
 {
